refactor: extract reaction.transition sub-module
