+++ src/expertsystem/amplitude/_graph_info.py
@@ -1,7 +1,11 @@
 from typing import Dict, List, Optional, Tuple, Union
 
-from expertsystem.reaction import InteractionProperties, ParticleWithSpin
-from expertsystem.reaction.particle import ParticleCollection, Spin
+from expertsystem.reaction import InteractionProperties
+from expertsystem.reaction.particle import (
+    ParticleCollection,
+    ParticleWithSpin,
+    Spin,
+)
 from expertsystem.reaction.topology import StateTransitionGraph, Topology
