+++ tests/unit/io/test_io.py
@@ -3,13 +3,14 @@ import json
 import pytest
 
 from expertsystem import io
-from expertsystem.reaction import (
-    Result,
+from expertsystem.reaction import Result
+from expertsystem.reaction.particle import Particle, ParticleCollection
+from expertsystem.reaction.topology import (
+    StateTransitionGraph,
+    Topology,
     create_isobar_topologies,
     create_n_body_topology,
 )
-from expertsystem.reaction.particle import Particle, ParticleCollection
-from expertsystem.reaction.topology import StateTransitionGraph, Topology
 
 
 def through_dict(instance):
